# coalgov-ai/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoalGov AI API",
    "description": "Mine governance pilot: inspections, compliance, production approvals, grievances and offline field reporting.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/mines": {
      "get": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Mines",
        "operationId": "mines_api_mines_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Add Mine",
        "operationId": "add_mine_api_mines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/inspections": {
      "get": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Inspections",
        "operationId": "inspections_api_inspections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Schedule",
        "operationId": "schedule_api_inspections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/inspections/{record_id}": {
      "put": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Update Inspection",
        "operationId": "update_inspection_api_inspections__record_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compliance": {
      "get": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Compliance",
        "operationId": "compliance_api_compliance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Add Compliance",
        "operationId": "add_compliance_api_compliance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/compliance/{record_id}": {
      "put": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Update Compliance",
        "operationId": "update_compliance_api_compliance__record_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/permits": {
      "get": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Permits",
        "operationId": "permits_api_permits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Add Permit",
        "operationId": "add_permit_api_permits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermitCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/permits/{record_id}": {
      "put": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Change Permit",
        "operationId": "change_permit_api_permits__record_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermitUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attendance": {
      "get": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Attendance",
        "operationId": "attendance_api_attendance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/attendance/check-in": {
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Check In",
        "operationId": "check_in_api_attendance_check_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/attendance/{record_id}/check-out": {
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Check Out",
        "operationId": "check_out_api_attendance__record_id__check_out_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets": {
      "get": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Assets",
        "operationId": "assets_api_assets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mine operations"
        ],
        "summary": "Add Asset",
        "operationId": "add_asset_api_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetCreate": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "name": {
            "type": "string",
            "maxLength": 140,
            "minLength": 3,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "maxLength": 60,
            "minLength": 2,
            "title": "Kind"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "name",
          "kind",
          "latitude",
          "longitude"
        ],
        "title": "AssetCreate"
      },
      "AttendanceInput": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id"
        ],
        "title": "AttendanceInput"
      },
...
      "ChecklistItem": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 250,
            "minLength": 2,
            "title": "Label"
          },
          "result": {
            "type": "string",
            "enum": [
              "pending",
              "pass",
              "fail",
              "na"
            ],
            "title": "Result",
            "default": "pending"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label"
        ],
        "title": "ChecklistItem"
      },
      "ComplianceCreate": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 3,
            "title": "Title"
          },
          "reference": {
            "type": "string",
            "maxLength": 220,
            "title": "Reference",
            "default": ""
          },
          "source_url": {
            "type": "string",
            "maxLength": 2000,
            "title": "Source Url",
            "default": ""
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "title",
          "owner_id",
          "due_at"
        ],
        "title": "ComplianceCreate"
      },
      "ComplianceUpdate": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "submitted",
              "compliant",
              "pending"
            ],
            "title": "Status"
          },
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "notes": {
            "type": "string",
            "maxLength": 10000,
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "status"
        ],
        "title": "ComplianceUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InspectionCreate": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "officer_id": {
            "type": "string",
            "title": "Officer Id"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Checklist"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "title",
          "officer_id",
          "scheduled_at",
          "checklist"
        ],
        "title": "InspectionCreate"
      },
      "InspectionUpdate": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "in_progress",
              "completed"
            ],
            "title": "Status"
          },
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Checklist"
          },
          "notes": {
            "type": "string",
            "maxLength": 10000,
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "status",
          "checklist"
        ],
        "title": "InspectionUpdate"
      },
...
      "MineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Name"
          },
          "subsidiary": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Subsidiary"
          },
          "state": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "State"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "subsidiary",
          "state",
          "latitude",
          "longitude"
        ],
        "title": "MineCreate"
      },
      "PermitCreate": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 3,
            "title": "Title"
          },
          "contractor_id": {
            "type": "string",
            "title": "Contractor Id"
          },
          "work_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Work Type"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "precautions": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 10,
            "title": "Precautions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "title",
          "contractor_id",
          "work_type",
          "starts_at",
          "expires_at",
          "precautions"
        ],
        "title": "PermitCreate"
      },
      "PermitUpdate": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "closed"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "status"
        ],
        "title": "PermitUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}